Implement most use cases of protocol 3.0
Complexity is already quite high, so some things were intentionally
left out (premium location, contract parking) of this initial
implementation to restrict the scope to the requirements of the initial
project (ECE/Easy-To-Park).

Although I originally wanted to adapt the facade pattern to the user
data as well, it turned out that a lot of specialized logic is
required to synchronize the data.
Since a dump of all user data needs to be sent and retrieved in
individual requests, the only way to achieve memory and - with regards
to object instantiation - CPU performance, was to use streaming and
enumerables, in turn making a single `ActiveRecord::Relation` necessary,
which was only possible to achieve with some carefully hand-crafted SQL.