Properly deal with duplicate card numbers
Since card numbers are not guaranteed to be unique anymore, we can't
just pass them downstream to the facade but need to do our own lookup
based on the cardkey/mediakey. Luckily, this has become clearer in
version 3.2 of the API and is specified as being supplied for each
request.