Refactor updating customer account mappings at ICA
The ICA synchronisation is messed up, due to bad scaling.
It fails with timeouts due to badly implemented queries.
Currently after a next RFID tag has been assigned to
customers, they are not marked as uploaded.
The issue should be fixed.
Furthermore the ICA customer_account_mapping-updating part
should be refactored to some extent in way, that it is not
affected by timeouts anymore.